fix: OpenJD architecture capability reporting on ARM
