+++ test/unit/startup/test_capabilities.py
@@ -3,10 +3,14 @@
 """Tests for the deadline_worker_agent.startup.capabilities module"""
 
 from typing import Any
-from pydantic import ValidationError
+from unittest.mock import MagicMock, patch
 import pytest
+import subprocess
+
+from pydantic import ValidationError
 
 from deadline_worker_agent.startup.capabilities import Capabilities
+from deadline_worker_agent.startup import capabilities as capabilities_mod
 
 
 @pytest.mark.parametrize(
@@ -195,3 +199,132 @@ def test_merge(
 
     # THEN
     assert result == expected_result
+
+
+@pytest.mark.parametrize(
+    argnames=("platform_machine", "expected_arch"),
+    argvalues=(
+        pytest.param("x86_64", "x86_64", id="intel-x86-64bit"),
+        pytest.param("amd64", "x86_64", id="amd-x86-64bit"),
+        pytest.param("arm64", "arm64", id="macos-arm"),
+        pytest.param("aarch64", "arm64", id="macos-arm"),
+    ),
+)
+def test_get_arch(
+    platform_machine: str,
+    expected_arch: str,
+) -> None:
+    """Tests that the _get_arch() function returns the correctly mapped value from
+    platform.machine()"""
+
+    # GIVEN
+    with patch.object(capabilities_mod.platform, "machine", return_value=platform_machine):
+        # WHEN
+        arch = capabilities_mod._get_arch()
+
+    # THEN
+    assert arch == expected_arch
+
+
+class TestGetGPUCount:
+    @patch.object(capabilities_mod.subprocess, "check_output")
+    def test_get_gpu_count(
+        self,
+        check_output_mock: MagicMock,
+    ) -> None:
+        """
+        Tests that the _get_gpu_count function returns the correct number of GPUs
+        """
+        # GIVEN
+        check_output_mock.return_value = b"2"
+
+        # WHEN
+        result = capabilities_mod._get_gpu_count()
+
+        # THEN
+        check_output_mock.assert_called_once_with(
+            ["nvidia-smi", "--query-gpu=count", "--format=csv,noheader"]
+        )
+        assert result == 2
+
+    @pytest.mark.parametrize(
+        ("exception", "expected_result"),
+        (
+            pytest.param(FileNotFoundError("nvidia-smi not found"), 0, id="FileNotFoundError"),
+            pytest.param(subprocess.CalledProcessError(1, "command"), 0, id="CalledProcessError"),
+        ),
+    )
+    @patch.object(capabilities_mod.subprocess, "check_output")
+    def test_get_gpu_count_nvidia_smi_error(
+        self,
+        check_output_mock: MagicMock,
+        exception: Exception,
+        expected_result: int,
+    ) -> None:
+        """
+        Tests that the _get_gpu_count function returns 0 when nvidia-smi is not found or fails
+        """
+        # GIVEN
+        check_output_mock.side_effect = exception
+
+        # WHEN
+        result = capabilities_mod._get_gpu_count()
+
+        # THEN
+        check_output_mock.assert_called_once_with(
+            ["nvidia-smi", "--query-gpu=count", "--format=csv,noheader"]
+        )
+
+        assert result == expected_result
+
+
+class TestGetGPUMemory:
+    @patch.object(capabilities_mod.subprocess, "check_output")
+    def test_get_gpu_memory(
+        self,
+        check_output_mock: MagicMock,
+    ) -> None:
+        """
+        Tests that the _get_gpu_memory function returns total memory
+        """
+        # GIVEN
+        check_output_mock.return_value = b"6800 MiB"
+
+        # WHEN
+        result = capabilities_mod._get_gpu_memory()
+
+        # THEN
+        check_output_mock.assert_called_once_with(
+            ["nvidia-smi", "--query-gpu=memory.total", "--format=csv,noheader"]
+        )
+        assert result == 6800
+
+    @pytest.mark.parametrize(
+        ("exception", "expected_result"),
+        (
+            pytest.param(FileNotFoundError("nvidia-smi not found"), 0, id="FileNotFoundError"),
+            pytest.param(subprocess.CalledProcessError(1, "command"), 0, id="CalledProcessError"),
+        ),
+    )
+    @patch.object(capabilities_mod.subprocess, "check_output")
+    def test_get_gpu_memory_nvidia_smi_error(
+        self,
+        check_output_mock: MagicMock,
+        exception: Exception,
+        expected_result: int,
+    ) -> None:
+        """
+        Tests that the _get_gpu_memory function returns 0 when nvidia-smi is not found or fails
+        """
+        # GIVEN
+        check_output_mock.side_effect = exception
+
+        # WHEN
+        result = capabilities_mod._get_gpu_memory()
+
+        # THEN
+        check_output_mock.assert_called_once_with(
+            ["nvidia-smi", "--query-gpu=memory.total", "--format=csv,noheader"]
+        )
+
+        assert result == expected_result
